Seguridad Social funcionando

+++ src/main/java/org/alfime/apoyo/domain/projection/EvaluacionIngresoCompleto.java
@@ -11,7 +11,7 @@ public interface EvaluacionIngresoCompleto {
 
     Ingreso getIngreso();
 
-    Double getMonto();
+    Long getMonto();
 
     Frecuencia getFrecuencia();
 }
